A pass at warning cleanup, including those shown by UBuntu 11, plus examples Makefile improvements. Makefile: add new targets: examples, cleanexamples, runexamples adapt/os/ocpios/linux/src/OcpiOsProcessManager.cxx: warnings, typos in comments adapt/os/ocpios/linux/src/OcpiOsSpinLock.cxx: warnings components/optical_flow_pyr_lk.rcc/optical_flow_pyr_lk.c: warnings components/specs/min_eigen_val_spec.xml: fixed spec name to be correct core/application/src/OcpiApplication.cxx: warnings core/container/hdl_container/bin/src/testRpl_main.cxx: warnings core/container/hdl_container/impl/src/XyzContainer.cxx: warnings core/container/interfaces/api/src/OcpiContainerInterface.cxx: warnings which found a bug! core/container/interfaces/api/src/OcpiContainerManager.cxx: warning core/container/ocl_container/impl/src/OcpiOclPlatformManager.cxx: print more device info core/dataplane/rdma_driver_interface/interfaces/include/DtHandshakeControl.h: warning found a bug! core/dataplane/rdma_drivers/sockets/src/DtSocketsXfer.cxx: warnings core/dataplane/rdma_smb/linux/src/OcpiPosixFileMappingServices.cxx: creator removes shm on destruction core/dataplane/transport/impl/include/OcpiPort.h: warning fix, also 32->64 bit. core/dataplane/transport/impl/src/OcpiPort.cxx: ditto core/library/src/OcpiComponentLibrary.cxx: improved error/exception message core/local/util/bin/src/fasttimed_main.cxx: warning core/local/util/ezxml/src/ezxml.c: warning core/local/util/fasttime/include/fasttime.h: warning core/local/util/property/src/OcpiUtilImplementation.cxx: warning found a bug! (in new not yet tested feature path) core/local/util/property/src/OcpiUtilPValue.cxx: formatting core/local/util/property/src/OcpiUtilValue.cxx: warning found a string-quoting bug core/local/util/vfs/src/OcpiUtilVfsUtil.cxx: warning hdl/primitives/Makefile: add "diff_cores" target to help when syncing upstream repo jkamd_env_linux_x86_64.sh: enable OpenCLb test/hdl/src/mem_probe.c: warning test/hdl/src/nft.c: warning test/hdl/src/swctl.c: warning tools/cdk/examples/bestfit_test/Makefile: update to use new example makefile scheme, include all test cases tools/cdk/examples/bestfit_test/README.txt: ditto tools/cdk/examples/bestfit_test/bestfit.cxx: return correct exit code tools/cdk/examples/bestfit_test/boston_small.jpg: removed tools/cdk/examples/bestfit_test/impl1.rcc/impl1.c: add required headers tools/cdk/examples/bestfit_test/impl2.rcc/impl2.c: ditto, plus warnings tools/cdk/examples/bestfit_test/impl3.rcc/impl3.c: ditto tools/cdk/examples/bestfit_test/runit.sh: removed tools/cdk/examples/canny/Makefile: makefile upgrade to new examples makefile scheme tools/cdk/examples/canny/canny.cxx: move some messages to stderr tools/cdk/examples/dds_ports/Makefile: makefile upgrade to new examples makefile scheme tools/cdk/examples/feature_detection/Makefile: ditto tools/cdk/examples/hello/Makefile: ditto tools/cdk/examples/hello/hello.cxx: add delay to give it enough time to work reliably tools/cdk/examples/image_filtering/Makefile: makefile upgrade to new examples makefile scheme tools/cdk/examples/image_filtering/runit: removed tools/cdk/examples/optical_flow/Makefile: makefile upgrade to new examples makefile scheme tools/cdk/examples/vsadd/Makefile: ditto tools/cdk/include/hdl/hdl-import-cores.mk: implement new diff_cores feature to help syncing upstream repo tools/cdk/include/hdl/hdl-import.mk: ditto tools/cdk/include/lib.mk: check for existence/installation of OpenCL and act accordingly tools/cdk/include/ocpisetup.mk: determine whether OpenCL is present, add linker features (-rpath etc.) tools/cdk/ocpidds/ocpidds/src/dataTypeTest.cxx: warning tools/cdk/ocpidds/ocpidds/src/ocpidds.cxx: warning tools/cdk/ocpigen/ocpigen/src/wip_HDL.cxx: warning tools/cdk/ocpigen/ocpigen/src/wip_parse.cxx: warning tools/cdk/ocpiocl/bin/src/ocpiocl_main.cxx: added test features to determine presense of OpenCL tools/local/tester/bin/src/ocpi-rcc-test_main.cxx: warnings

+++ core/dataplane/rdma_driver_interface/interfaces/include/DtHandshakeControl.h
@@ -153,7 +153,7 @@ namespace DataTransfer {
       RequestTypes            request;
       OCPI::OS::int32_t              error_code;
       OCPI::OS::uint32_t              returnMailboxId;
-      OCPI::OS::int32_t              return_offset;
+      OCPI::OS::int64_t              return_offset;
       OCPI::OS::uint32_t              return_size;
     };
